requirements.txt
numpy>=1.11.0
Pillow>=4.2.1 
spacy>=1.8.2
tweepy>=3.5.0
PyYAML>=3.12
praw>=5.0.1
pronouncing>=0.1.5
